// apps/api/src/setting/dto/update-company.dto.ts
export class UpdateCompanyDto {
  companyName?: string;
  logo?: string;
  stamp?: string;
  address?: string;
  phone?: string;
  email?: string;
  website?: string;
  taxCode?: string;
  bankName?: string;
  bankAccountNumber?: string;
  bankAccountHolder?: string;
  currency?: string;
  currencySymbol?: string;
  timezone?: string;
}